revert back commented out features

+++ src/args/logout/mod.ts
@@ -1,4 +1,9 @@
-import { defineCommand } from "@reliverse/prompts";
+import {
+  defineCommand,
+  deleteLastLine,
+  msg,
+  togglePrompt,
+} from "@reliverse/prompts";
 import relinka from "@reliverse/relinka";
 
 import { isConfigExists } from "~/utils/config.js";
@@ -8,12 +13,12 @@ import { deleteConfig } from "./impl.js";
 export default defineCommand({
   meta: {
     name: "logout",
-    description: "Forgets your credentials",
+    description: "Ask Reliverse to forget your data",
   },
   args: {
     dev: {
       type: "boolean",
-      description: "Runs the CLI in dev mode",
+      description: "Run the CLI in dev mode",
     },
   },
   run: async ({ args }) => {
@@ -25,8 +30,21 @@ export default defineCommand({
         : relinka.info("Try `reliverse login` cmd.");
       process.exit(0);
     }
-    await deleteConfig();
-    relinka.success("`You're logged out now!` 👋");
+    const danger = await togglePrompt({
+      title:
+        "Are you sure you want to log out? Reliverse will delete its local memory. This action cannot be undone.",
+      titleColor: "redBright",
+      options: ["Yes", "No"],
+      defaultValue: "No",
+    });
+    if (danger) {
+      await deleteConfig();
+      deleteLastLine();
+      msg({
+        type: "M_MIDDLE",
+      });
+      relinka.success("`You're logged out now!` 👋");
+    }
     process.exit(0);
   },
 });
